+++ common/src/main/java/com/liskovsoft/smartyoutubetv2/common/app/presenters/SplashPresenter.java
@@ -31,6 +31,7 @@ import com.liskovsoft.smartyoutubetv2.common.utils.IntentExtractor;
 import com.liskovsoft.smartyoutubetv2.common.utils.SimpleEditDialog;
 import com.liskovsoft.smartyoutubetv2.common.utils.Utils;
 import com.liskovsoft.leanbackassistant.misc.ServiceManagerProvider;
+import com.liskovsoft.plexapi.PlexServiceManager;
 import com.liskovsoft.smartyoutubetv2.common.misc.MediaSourceRegistry;
 
 import java.util.ArrayList;
@@ -93,6 +94,7 @@ public class SplashPresenter extends BasePresenter<SplashView> {
             // Wire ATV leanbackassistant to MediaSourceRegistry (avoids circular Gradle dep).
             ServiceManagerProvider.init(MediaSourceRegistry.getServiceManager());
             initGlobalPrefs();
+            PlexServiceManager.init(getContext());
             initProxy();
             initVideoStateService();
             initStreamReminderService();